Use melpa-stable for selectrum, consult, prescient due to instability
It is really annoying when these packages break because I feel like I
depend on them the most for my workflow.
I know selectrum and prescient have broken for me recently. Using
stable for consult as well because I know that's also a newer package
under development, and it interacts with the others (albeit indirectly).